separate estimators from GFNModules
