+++ src/gfn/gym/helpers/box_utils.py
@@ -8,6 +8,7 @@ import torch.nn as nn
 from torch import Size, Tensor
 from torch.distributions import Beta, Categorical, Distribution, MixtureSameFamily
 
+from gfn.estimators import Estimator
 from gfn.gym import Box
 from gfn.modules import GFNModule
 from gfn.states import States
@@ -868,7 +869,7 @@ class BoxStateFlowModule(MLP):
         return out
 
 
-class BoxPBUniform(nn.Module):
+class BoxPBUniform(GFNModule):
     """A uniform backward policy for the Box environment.
 
     This module returns `(1, 1, 1)` for all states. Used with `QuarterCircle`,
@@ -936,7 +937,7 @@ def split_PF_module_output(
     return (exit_probability, mixture_logits, alpha_theta, beta_theta, alpha_r, beta_r)
 
 
-class BoxPFEstimator(GFNModule):
+class BoxPFEstimator(Estimator):
     r"""Estimator for `P_F` for the Box environment.
 
     This estimator uses the `DistributionWrapper` distribution.
@@ -961,7 +962,7 @@ class BoxPFEstimator(GFNModule):
     def __init__(
         self,
         env: Box,
-        module: nn.Module,
+        module: GFNModule,
         n_components_s0: int,
         n_components: int,
         min_concentration: float = 0.1,
@@ -1060,7 +1061,7 @@ class BoxPFEstimator(GFNModule):
         )
 
 
-class BoxPBEstimator(GFNModule):
+class BoxPBEstimator(Estimator):
     r"""Estimator for `P_B` for the Box environment.
 
     This estimator uses the `QuarterCircle(northeastern=False)` distribution.
@@ -1080,7 +1081,7 @@ class BoxPBEstimator(GFNModule):
     def __init__(
         self,
         env: Box,
-        module: nn.Module,
+        module: GFNModule,
         n_components: int,
         min_concentration: float = 0.1,
         max_concentration: float = 2.0,
