+++ src/gfn/samplers.py
@@ -5,7 +5,7 @@ import torch
 from gfn.actions import Actions
 from gfn.containers import Trajectories
 from gfn.env import Env
-from gfn.modules import GFNModule
+from gfn.estimators import Estimator
 from gfn.states import GraphStates, States
 from gfn.utils.common import ensure_same_device
 from gfn.utils.graphs import graph_states_share_storage
@@ -28,7 +28,7 @@ class Sampler:
             probability distributions.
     """
 
-    def __init__(self, estimator: GFNModule) -> None:
+    def __init__(self, estimator: Estimator) -> None:
         """Initializes a Sampler with a PolicyEstimator.
 
         Args:
@@ -330,7 +330,7 @@ class LocalSearchSampler(Sampler):
         backward_sampler: A Sampler instance with the backward PolicyEstimator.
     """
 
-    def __init__(self, pf_estimator: GFNModule, pb_estimator: GFNModule):
+    def __init__(self, pf_estimator: Estimator, pb_estimator: Estimator):
         """Initializes a LocalSearchSampler with forward and backward estimators.
 
         Args:
