+++ testing/test_gflownet.py
@@ -1,9 +1,9 @@
 from gfn.containers import StatesContainer, Trajectories
 from gfn.containers.base import Container
+from gfn.estimators import DiscretePolicyEstimator
 from gfn.gflownet import FMGFlowNet, TBGFlowNet
 from gfn.gym import Box, HyperGrid
 from gfn.gym.helpers.box_utils import BoxPBEstimator, BoxPBMLP, BoxPFEstimator, BoxPFMLP
-from gfn.modules import DiscretePolicyEstimator
 from gfn.preprocessors import KHotPreprocessor
 from gfn.states import DiscreteStates
 from gfn.utils.modules import MLP
