+++ tutorials/examples/train_line.py
@@ -7,6 +7,7 @@ from torch.distributions import Distribution, Normal  # TODO: extend to Beta
 from torch.distributions.independent import Independent
 from tqdm import trange
 
+from gfn.estimators import Estimator
 from gfn.gflownet import TBGFlowNet  # TODO: Extend to SubTBGFlowNet
 from gfn.gym.line import Line
 from gfn.modules import GFNModule
@@ -168,10 +169,10 @@ class GaussianStepMLP(MLP):
         return out
 
 
-class StepEstimator(GFNModule):
+class StepEstimator(Estimator):
     """Estimator for PF and PB of the Line environment."""
 
-    def __init__(self, env: Line, module: torch.nn.Module, backward: bool):
+    def __init__(self, env: Line, module: GFNModule, backward: bool):
         super().__init__(module, is_backward=backward)
         self.backward = backward
         self.n_steps_per_trajectory = env.n_steps_per_trajectory
